requirements.txt
flask
gunicorn
python-telegram-bot==20.7
aiohttp
